requirements.txt
asgiref==3.7.2
boto3==1.34.1
botocore==1.34.1
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.23
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-storages==1.14.2
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
stripe==7.8.2
typing_extensions==4.9.0
urllib3==2.0.7
